+++ packages/common/core-interfaces/src/brandedType.ts
@@ -82,7 +82,7 @@
  * }
  * ```
  *
- * @beta
+ * @public
  */
 export declare class BrandedType<out Brand> {
 	/**