+++ packages/framework/fluid-framework/src/index.ts
@@ -33,6 +33,7 @@ export type {
 	MemberChangedListener,
 	Myself,
 } from "@fluidframework/fluid-static";
+export { getPresence, getPresenceAlpha } from "@fluidframework/fluid-static/internal";
 export type { SharedObjectKind } from "@fluidframework/shared-object-base";
 export type {
 	IErrorBase,
@@ -101,10 +102,10 @@ import {
  * A hierarchical data structure for collaboratively editing strongly typed JSON-like trees
  * of objects, arrays, and other data types.
  * @privateRemarks
- * Here we reexport SharedTree, but with the `@alpha` types (`ISharedObjectKind`) removed, just keeping the `SharedObjectKind`.
+ * Here we reexport SharedTree, but with the `@legacy` types (`ISharedObjectKind`) removed, just keeping the `SharedObjectKind`.
  * Doing this requires creating this new typed export rather than relying on a reexport directly from the tree package.
  * The tree package itself does not do this because it's API needs to be usable from the encapsulated API which requires `ISharedObjectKind`.
- * This package however is not intended for use by users of the encapsulated API, and therefor it can discard that interface.
+ * This package however is not intended for use by users of the encapsulated API, and therefore it can discard that interface.
  * @public
  */
 export const SharedTree: SharedObjectKind<ITree> = OriginalSharedTree;
@@ -188,6 +189,7 @@ export { SharedString } from "@fluidframework/sequence/internal";
 export type {
 	ISharedObject,
 	ISharedObjectEvents,
+	ISharedObjectKind,
 } from "@fluidframework/shared-object-base/internal";
 
 export type {